requirements.txt
Django==4.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.5
